Start refactoring of events
This includes:
* migration to WeakMap for private members
* factoring out every object into its own file

Since our migration to the vm module, this also requires the introduction of a MultihandleWeakMap to account for the multiple window handles.
The MultihandleWeakMap allows aliasing of handles, in the normal case, it'll just redirect to the underlying WeakMap store, however if it doesn't find the handle there, it'll look into an alias map to check if there're other aliases to the object. This allows us to work with the window normally.